Fix debug-on-error
This was a pain to track down.  I turn it off early in .emacs, but
something was turning it back on.  Turns out it was nxhtml.  A glance
at their bzr repository lp:nxhtml shows that this will be "fixed" in
2.09.  I say "fixed" because if your emacs happens to be in
C:\EmacsW32, it will still be turned on by default.