Ignore unsafe variables with desktop & daemon mode
When a daemonized emacs is initially started and desktop-save-mode
loads a desktop it may prompt the user on a detached tty to ask what
to do about potentially unsafe local variables and autosave files.  I
already implemented a fix for autosave files using advice and have
added support for automatically ignoring potentially unsafe local
variables in the same function.