Delay load desktop to avoid hanging daemon mode
Load it when Emacs is started without the --daemon switch, or when an
emacsclient opens a new frame or switches buffers.  In these cases we do
not have to worry about questions being asked without an interface to
answer them.  One oddity is that if emacsclient visits a file that is in
the desktop, a spurious question may be presented.