Activate ws-butler in pre-existing buffers
Desktop can load and activate buffers that should be in ws-butler-mode
before ws-butler is loaded, due to the fact ws-butler mode cannot be
activated until a color or grayscale frame or TTY has been attached.
So, this will look at all existing buffers when ws-butler loads and
activate it in the ones that should have it.