Use explicit load-paths for packages in site-lisp
This is in preparation for loading different versions of packages
from different subdirectories depending upon the version of emacs
being run.

Blame it on magit. I need one of three different versions of it
and it's support libraries depending upon what version of emacs
is on the system I am using at the moment.