Only create a custom js-dict for AC when necessary
The latest version has a js-dict, but no javascript-dict. This was causing
the auto-complete :config logic to fail. So only copy the javascript-dict
to a custom js-dict if there is no js-dict.

I am, however, retaining the copy logic in case it becomes necessary to
support multiple versions of AC for various versions of emacs or in case
it becomes necessary to create a custom javascript-dict from the now extant
js-dict for older versions of emacs.