Improve my Drew Adam's update.sh script
The beginning of the script downloads a new emacs-init.el and adjusts
paths in it to point to where I store Drew's code on my system. Be
careful not to accidentally replace it with a freshly-downloaded,
unmodified version.

Also check the features list of all of his files. The list in
emacs-init.el does a fairly good job, but occassionally misses
something that's listed elsewhere.