Set the minibuffer position based upon the `window-system'
X Windows in WSL (Ubuntu on Windows) with a Microsoft Windows X Server
insists on placing the minibuffer at the top of the screen and it
gets hidden behind other windows.  So place it at the bottom of the
screen, just above the taskbar.

When running natively in Microsoft Windows, when active, move the
minibuffer near point.  This does not work in X Windows on MS Windows
because the minibuffer does not get raised and is hidden behind the
window that has focus.